+++ src/main/java/project/flipnote/user/model/UserRegisterResponse.java
@@ -1,0 +1,10 @@
+package project.flipnote.user.model;
+
+public record UserRegisterResponse(
+	Long userId
+) {
+
+	public static UserRegisterResponse from(Long userId) {
+		return new UserRegisterResponse(userId);
+	}
+}
